+++ packages/@livestore/adapter-cloudflare/src/make-adapter.ts
@@ -109,7 +109,7 @@ export const makeAdapter =
               pull: ({ cursor }) => syncProcessor.pull({ cursor }),
               push: (batch) =>
                 syncProcessor.push(
-                  batch.map((item) => new LiveStoreEvent.EncodedWithMeta(item)),
+                  batch.map((item) => new LiveStoreEvent.Client.EncodedWithMeta(item)),
                   { waitForProcessing: true },
                 ),
             },