+++ packages/@livestore/adapter-expo/src/make-sqlite-db.ts
@@ -76,7 +76,7 @@ const makeSqliteDb_ = <TMetadata extends Metadata>({
     _tag: 'SqliteDb',
     debug: {
       // Setting initially to root but will be set to correct value shortly after
-      head: EventSequenceNumber.ROOT,
+      head: EventSequenceNumber.Client.ROOT,
     },
     prepare: (queryStr) => {
       try {