+++ packages/@livestore/adapter-node/src/client-session/adapter.ts
@@ -338,7 +338,7 @@ const makeLocalLeaderThread = ({
             pull: ({ cursor }) => syncProcessor.pull({ cursor }),
             push: (batch) =>
               syncProcessor.push(
-                batch.map((item) => new LiveStoreEvent.EncodedWithMeta(item)),
+                batch.map((item) => new LiveStoreEvent.Client.EncodedWithMeta(item)),
                 { waitForProcessing: true },
               ),
           },