+++ packages/@livestore/adapter-node/src/make-leader-worker.ts
@@ -74,7 +74,7 @@ export const makeWorkerEffect = (options: WorkerOptions) => {
     PushToLeader: ({ batch }) =>
       Effect.andThen(LeaderThreadCtx, (_) =>
         _.syncProcessor.push(
-          batch.map((item) => new LiveStoreEvent.EncodedWithMeta(item)),
+          batch.map((item) => new LiveStoreEvent.Client.EncodedWithMeta(item)),
           // We'll wait in order to keep back pressure on the client session
           { waitForProcessing: true },
         ),