+++ packages/@livestore/adapter-web/src/in-memory/in-memory-adapter.ts
@@ -236,7 +236,7 @@ const makeLeaderThread = ({
           pull: ({ cursor }) => syncProcessor.pull({ cursor }),
           push: (batch) =>
             syncProcessor.push(
-              batch.map((item) => new LiveStoreEvent.EncodedWithMeta(item)),
+              batch.map((item) => new LiveStoreEvent.Client.EncodedWithMeta(item)),
               { waitForProcessing: true },
             ),
         },